requirements.txt
django
channels["daphne"]
channels-redis
djangorestframework
PyJWT
django-cors-headers
django-phonenumber-field[phonenumbers]
django-twilio